Update keypad_m_vid.ino
Updated the data type from int to long, now it can store 32 bits (4 bytes), from -2,147,483,648 to 2,147,483,647.

+++ keypad_m_vid.ino
@@ -28,11 +28,11 @@ void setup() {
     Serial.begin(9600);
     
 }
-int getKeypadIntegerMulti()
+long getKeypadIntegerMulti()
 {
   
-  int value = 0;                                // the number accumulator
-  int keyvalue;                                     // the key pressed at current moment
+  long value = 0;                                // the number accumulator
+  long keyvalue;                                     // the key pressed at current moment
   int isnum;
   Serial.println("Enter the digits,press any non-digit to end ");
   Serial.print("You have typed: ");
